新增签名保存

+++ src/main/java/com/houses/service/impl/IHouseServiceImpl.java
@@ -194,15 +194,17 @@ public class IHouseServiceImpl implements IHouseService {
                 File fullImage = new File(currItem.getFullItemExampleImage());
                 //所有土拍你都要复制，然后删除原有图片
                 //图片路径不一致，复制图片
-                if (!currItem.getFullItemExampleImage().equals(UPLOAD_PATH + File.separator + fullImage.getName())) {
-                    FileUtils.copyFile(fullImage, new File(UPLOAD_PATH + File.separator + fullImage.getName()));
-                    currItem.setFullItemExampleImage(UPLOAD_PATH + File.separator + fullImage.getName());
-                }else {
-                    //路径一致，换个马甲，删除原来的图片
-                    long time = System.currentTimeMillis();
-                    FileUtils.copyFile(fullImage, new File(UPLOAD_PATH + File.separator + time + IMAGE_SUFFIX));
-                    FileUtils.deleteQuietly(fullImage);
-                    currItem.setFullItemExampleImage(UPLOAD_PATH + File.separator + time + IMAGE_SUFFIX);
+                if (fullImage.isFile()) {
+                    if (!currItem.getFullItemExampleImage().equals(UPLOAD_PATH + File.separator + fullImage.getName())) {
+                        FileUtils.copyFile(fullImage, new File(UPLOAD_PATH + File.separator + fullImage.getName()));
+                        currItem.setFullItemExampleImage(UPLOAD_PATH + File.separator + fullImage.getName());
+                    } else {
+                        //路径一致，换个马甲，删除原来的图片
+                        long time = System.currentTimeMillis();
+                        FileUtils.copyFile(fullImage, new File(UPLOAD_PATH + File.separator + time + IMAGE_SUFFIX));
+                        FileUtils.deleteQuietly(fullImage);
+                        currItem.setFullItemExampleImage(UPLOAD_PATH + File.separator + time + IMAGE_SUFFIX);
+                    }
                 }
             } catch (IOException e) {
                 e.printStackTrace();
@@ -233,14 +235,16 @@ public class IHouseServiceImpl implements IHouseService {
                 currCrack.setItemId(currItem.getId());
                 File exampleImage = new File(currCrack.getExampleImage());
                 try {
-                    if (!currCrack.getExampleImage().equals(UPLOAD_PATH + File.separator + exampleImage.getName())) {
-                        FileUtils.copyFile(exampleImage, new File(UPLOAD_PATH + File.separator + exampleImage.getName()));
-                        currCrack.setExampleImage(UPLOAD_PATH + File.separator + exampleImage.getName());
-                    }else {
-                        long time = System.currentTimeMillis();
-                        FileUtils.copyFile(exampleImage, new File(UPLOAD_PATH + File.separator + time + IMAGE_SUFFIX));
-                        FileUtils.deleteQuietly(exampleImage);
-                        currCrack.setExampleImage(UPLOAD_PATH + File.separator + time + IMAGE_SUFFIX);
+                    if (exampleImage.isFile()) {
+                        if (!currCrack.getExampleImage().equals(UPLOAD_PATH + File.separator + exampleImage.getName())) {
+                            FileUtils.copyFile(exampleImage, new File(UPLOAD_PATH + File.separator + exampleImage.getName()));
+                            currCrack.setExampleImage(UPLOAD_PATH + File.separator + exampleImage.getName());
+                        } else {
+                            long time = System.currentTimeMillis();
+                            FileUtils.copyFile(exampleImage, new File(UPLOAD_PATH + File.separator + time + IMAGE_SUFFIX));
+                            FileUtils.deleteQuietly(exampleImage);
+                            currCrack.setExampleImage(UPLOAD_PATH + File.separator + time + IMAGE_SUFFIX);
+                        }
                     }
                 } catch (IOException e) {
                     e.printStackTrace();
